getCRUCLdata: Download and Create a Data Frame or Raster Stack Object of CRU CL v. 2.0 Climatology Variables

The getCRUCLdata package provides functions that automate downloading and importing CRU CL v. 2.0 climatology data, facilitates the calculation of minimum temperature and maximum temperature, converts elevation from kilometres to metres and formats the data into a tidy data frame or a list of stack objects for use. CRU CL v. 2.0 data are a gridded climatology of 1961-1990 monthly means released in 2002 and cover all land areas (excluding Antarctica) at 10 arcminutes (0.1666667 degree) resolution. For more information see the description of the data provided by the University of East Anglia Climate Research Unit (CRU), https://crudata.uea.ac.uk/cru/data/hrg/tmc/readme.txt.

References

https://crudata.uea.ac.uk/cru/data/hrg/tmc/

New, M., Lister, D., Hulme, M. and Makin, I., 2002: A high-resolution data set of surface climate over global land areas. Climate Research 21:1-25 (abstract, paper)

See also