getCRUdata: Download and Create a Data Frame or Raster Stack Object of CRU CL2.0 Climatology Variables

The getCRUdata package provides functions that automate downloading and importing CRU CL2.0 climatology data into R, facilitates the calculation of minimum temperature and maximum temperature, converts elevation from kilometres to metres and formats the data into a tidy data frame or a list of raster stack objects for use in an R session. CRU CL2.0 data are a gridded climatology of 1961-1990 monthly means released in 2002 and cover all land areas (excluding Antarctica) at 10 arc-minute resolution. For more information see the description of the data provided by the University of East Anglia Climate Research Unit (CRU), https://crudata.uea.ac.uk/cru/data/hrg/tmc/readme.txt.

References

https://crudata.uea.ac.uk/cru/data/hrg/tmc/

New, M., Lister, D., Hulme, M. and Makin, I., 2002: A high-resolution data set of surface climate over global land areas. Climate Research 21:1-25 (https://crudata.uea.ac.uk/cru/data/hrg/tmc/, http://www.int-res.com/articles/cr2002/21/c021p001.pdf)

See also

create_CRU_df

create_CRU_stack